+++ src/DemoFile/DemoParser.cs
@@ -16,7 +16,7 @@ public sealed partial class DemoParser
     private readonly Dictionary<DemoTick, long> _keyTickPositions = new();
     private readonly PriorityQueue<QueuedPacket, (int, int)> _packetQueue = new(128);
     private readonly PriorityQueue<ITickTimer, uint> _serverTickTimers = new();
-    private readonly Source1GameEvents _source1GameEvents = new();
+    private readonly Source1GameEvents _source1GameEvents;
     private DemoEvents _demoEvents;
     private EntityEvents _entityEvents;
     private GameEvents _gameEvents;
@@ -33,8 +33,16 @@ public sealed partial class DemoParser
     /// </remarks>
     public Action<DemoProgressEvent>? OnProgress;
 
+    /// <summary>
+    /// Event fired when the current demo command has finished (e.g, just before <see cref="MoveNextAsync"/> returns).
+    /// Reset to <c>null</c> just before it is invoked.
+    /// </summary>
+    public Action? OnCommandFinish;
+
     public DemoParser()
     {
+        _source1GameEvents = new Source1GameEvents(this);
+
         _stream = null!;
 
         _demoEvents.DemoFileHeader += msg => { FileHeader = msg; };
@@ -257,15 +265,25 @@ public sealed partial class DemoParser
         var buf = await ReadExactBytesAsync((int)cmd.Size, cancellationToken).ConfigureAwait(false);
 
         // todo: disable seeking until command is over!
+        bool canContinue;
         if (isCompressed)
         {
             using var decompressed = Snappy.DecompressToMemory(buf);
-            return _demoEvents.ReadDemoCommand(msgType, decompressed.Memory.Span);
+            canContinue = _demoEvents.ReadDemoCommand(msgType, decompressed.Memory.Span);
         }
         else
         {
-            return _demoEvents.ReadDemoCommand(msgType, buf);
+            canContinue = _demoEvents.ReadDemoCommand(msgType, buf);
         }
+
+        if (OnCommandFinish is { } onCommandFinish)
+        {
+            // Reset to null before invoking to allow any callbacks to re-register
+            OnCommandFinish = null;
+            onCommandFinish();
+        }
+
+        return canContinue;
     }
 
     private async ValueTask ReadFileInfo(CancellationToken cancellationToken)